feat(llm): rename dynamicContent file in useDynamicContent
