+++ apps/ledger-live-mobile/src/components/Carousel/CarouselCard.tsx
@@ -3,7 +3,7 @@ import { Linking } from "react-native";
 import { Flex, FullBackgroundCard } from "@ledgerhq/native-ui";
 import { useTheme } from "styled-components/native";
 import { WalletContentCard } from "~/dynamicContent/types";
-import useDynamicContent from "~/dynamicContent/dynamicContent";
+import useDynamicContent from "~/dynamicContent/useDynamicContent";
 import ForceTheme from "../theme/ForceTheme";
 
 type CarouselCardProps = {
