+++ apps/ledger-live-mobile/src/components/RootNavigator/ExploreTabNavigator.tsx
@@ -12,7 +12,7 @@ import Learn from "~/screens/Learn/learn";
 import ExploreTabNavigatorTabBar from "../ExploreTab/ExploreTabNavigatorTabBar";
 import ExploreTabNavigatorTabBarDisabled from "../ExploreTab/ExploreTabNavigatorTabBarDisabled";
 import { useIsNewsfeedAvailable } from "~/hooks/newsfeed/useIsNewsfeedAvailable";
-import useDynamicContent from "~/dynamicContent/dynamicContent";
+import useDynamicContent from "~/dynamicContent/useDynamicContent";
 import { NavigationHeaderBackButton } from "../NavigationHeaderBackButton";
 
 const ExploreTab = createMaterialTopTabNavigator<ExploreTabNavigatorStackParamList>();
