+++ apps/ledger-live-mobile/src/dynamicContent/ContentCardsLocation.tsx
@@ -1,12 +1,16 @@
-import React, { useCallback } from "react";
+import React from "react";
 import { FlatList, ListRenderItem } from "react-native";
 import { Flex } from "@ledgerhq/native-ui";
 import type { FlexBoxProps } from "@ledgerhq/native-ui/components/Layout/Flex/index";
-import useDynamicContent from "./dynamicContent";
+import useDynamicContent from "./useDynamicContent";
 import { BrazeContentCard, CategoryContentCard, ContentCardLocation } from "./types";
 import ContentCardsCategory from "./ContentCardsCategory";
 import { FeatureToggle } from "@ledgerhq/live-config/lib/featureFlags/index";
 import { filterCategoriesByLocation, formatCategories } from "./utils";
+
+const Category: ListRenderItem<CategoriesWithCards> = ({ item }) => (
+  <ContentCardsCategory category={item.category} categoryContentCards={item.cards} />
+);
 
 type Props = FlexBoxProps & {
   locationId: ContentCardLocation;
@@ -20,14 +24,6 @@ type CategoriesWithCards = {
 const ContentCardsLocation = ({ locationId, ...containerProps }: Props) => {
   const { categoriesCards, mobileCards } = useDynamicContent();
 
-  // TODO : extract in a component above
-  const renderCategory: ListRenderItem<CategoriesWithCards> = useCallback(
-    ({ item }) => (
-      <ContentCardsCategory category={item.category} categoryContentCards={item.cards} />
-    ),
-    [],
-  );
-
   const categoriesToDisplay = filterCategoriesByLocation(categoriesCards, locationId);
 
   if (categoriesToDisplay.length === 0) return null;
@@ -38,7 +34,7 @@ const ContentCardsLocation = ({ locationId, ...containerProps }: Props) => {
     <Flex {...containerProps}>
       <FlatList
         data={categoriesFormatted}
-        renderItem={renderCategory}
+        renderItem={Category}
         keyExtractor={(item: CategoriesWithCards) => item.category.id}
         ItemSeparatorComponent={() => <Flex height={8} />}
       />
