+++ apps/ledger-live-mobile/src/dynamicContent/useDynamicContent.tsx
@@ -16,7 +16,6 @@ import { track } from "../analytics";
 import { setDismissedDynamicCards } from "../actions/settings";
 import { setDynamicContentMobileCards } from "~/actions/dynamicContent";
 
-// TODO : rename file => use...
 const useDynamicContent = () => {
   const dispatch = useDispatch();
   const { logClickCard, logDismissCard, logImpressionCard, refreshDynamicContent } =
