+++ apps/ledger-live-mobile/src/screens/Learn/learn.tsx
@@ -7,7 +7,7 @@ import styled, { useTheme } from "styled-components/native";
 import { TrackScreen } from "~/analytics";
 import Skeleton from "~/components/Skeleton";
 import { ScreenName } from "~/const";
-import useDynamicContent from "~/dynamicContent/dynamicContent";
+import useDynamicContent from "~/dynamicContent/useDynamicContent";
 import { LearnContentCard } from "~/dynamicContent/types";
 
 const keyExtractor = (item: LearnContentCard) => item.id;
