+++ apps/ledger-live-mobile/src/screens/Portfolio/index.tsx
@@ -43,7 +43,7 @@ import { WalletTabNavigatorStackParamList } from "~/components/RootNavigator/typ
 import AddAccountsModal from "../AddAccounts/AddAccountsModal";
 import CollapsibleHeaderFlatList from "~/components/WalletTab/CollapsibleHeaderFlatList";
 import globalSyncRefreshControl from "~/components/globalSyncRefreshControl";
-import useDynamicContent from "~/dynamicContent/dynamicContent";
+import useDynamicContent from "~/dynamicContent/useDynamicContent";
 import PortfolioOperationsHistorySection from "./PortfolioOperationsHistorySection";
 import PortfolioGraphCard from "./PortfolioGraphCard";
 import {
@@ -166,37 +166,37 @@ function PortfolioScreen({ navigation }: NavigationProps) {
       ) : null,
       ...(showAssets && isAWalletCardDisplayed
         ? [
-          <Box background={colors.background.main} key="CarouselTitle">
-            <SectionContainer px={0} minHeight={240} isFirst>
-              <SectionTitle
-                title={t("portfolio.carousel.title")}
-                containerProps={{ mb: 7, mx: 6 }}
-              />
-              <Carousel />
-            </SectionContainer>
-          </Box>,
-        ]
+            <Box background={colors.background.main} key="CarouselTitle">
+              <SectionContainer px={0} minHeight={240} isFirst>
+                <SectionTitle
+                  title={t("portfolio.carousel.title")}
+                  containerProps={{ mb: 7, mx: 6 }}
+                />
+                <Carousel />
+              </SectionContainer>
+            </Box>,
+          ]
         : []),
       ...(showAssets
         ? [
-          <SectionContainer px={6} isFirst={!isAWalletCardDisplayed} key="AllocationsSection">
-            <SectionTitle title={t("analytics.allocation.title")} />
-            <Flex minHeight={94}>
-              <AllocationsSection />
-            </Flex>
-          </SectionContainer>,
-          <SectionContainer px={6} key="PortfolioOperationsHistorySection">
-            <SectionTitle title={t("analytics.operations.title")} />
-            <PortfolioOperationsHistorySection />
-          </SectionContainer>,
-        ]
+            <SectionContainer px={6} isFirst={!isAWalletCardDisplayed} key="AllocationsSection">
+              <SectionTitle title={t("analytics.allocation.title")} />
+              <Flex minHeight={94}>
+                <AllocationsSection />
+              </Flex>
+            </SectionContainer>,
+            <SectionContainer px={6} key="PortfolioOperationsHistorySection">
+              <SectionTitle title={t("analytics.operations.title")} />
+              <PortfolioOperationsHistorySection />
+            </SectionContainer>,
+          ]
         : [
-          // If the user has no accounts we display an empty state
-          <Flex flexDirection="column" mt={30} mx={6} key="PortfolioEmptyState">
-            <RecoverBanner />
-            <PortfolioEmptyState openAddAccountModal={openAddModal} />
-          </Flex>,
-        ]),
+            // If the user has no accounts we display an empty state
+            <Flex flexDirection="column" mt={30} mx={6} key="PortfolioEmptyState">
+              <RecoverBanner />
+              <PortfolioEmptyState openAddAccountModal={openAddModal} />
+            </Flex>,
+          ]),
     ],
     [
       onBackFromUpdate,
